Move `accel` implementation fully to `cuml.accel` (#6401)
Previously we'd added a shim to import the implementation from `cuml.experimental.accel` into `cuml.accel`. This PR finishes the move, moving everything out of `experimental` and into `cuml.accel` proper.

Authors:
  - Jim Crist-Harif (https://github.com/jcrist)

Approvers:
  - Gil Forsyth (https://github.com/gforsyth)
  - Simon Adorf (https://github.com/csadorf)

URL: https://github.com/rapidsai/cuml/pull/6401
