+++ ci/run_cuml_singlegpu_accel_pytests.sh
@@ -2,6 +2,6 @@
 # Copyright (c) 2024-2025, NVIDIA CORPORATION.
 
 # Support invoking run_cuml_singlegpu_pytests.sh outside the script directory
-cd "$(dirname "$(realpath "${BASH_SOURCE[0]}")")"/../python/cuml/cuml/tests/experimental/accel
+cd "$(dirname "$(realpath "${BASH_SOURCE[0]}")")"/../python/cuml/cuml/tests/accel
 
 python -m pytest -p cuml.accel --cache-clear "$@" .
