+++ ci/test_python_singlegpu.sh
@@ -25,7 +25,7 @@ rapids-logger "pytest cuml single GPU"
   --numprocesses=8 \
   --dist=worksteal \
   --junitxml="${RAPIDS_TESTS_DIR}/junit-cuml-accel.xml" \
-  --cov-config=../../../../.coveragerc \
+  --cov-config=../../../.coveragerc \
   --cov=cuml \
   --cov-report=xml:"${RAPIDS_COVERAGE_DIR}/cuml-accel-coverage.xml"
