+++ python/cuml/cuml/accel/estimator_proxy.py
@@ -14,16 +14,14 @@
 # limitations under the License.
 #
 
-
 import inspect
 import sys
 import types
+from typing import Optional, Tuple, Dict, Any, Type, List
 
-from cuml.internals.mem_type import MemoryType
 from cuml.internals import logger
 from cuml.internals.global_settings import GlobalSettings
 from cuml.internals.safe_imports import gpu_only_import, cpu_only_import
-from typing import Optional, Tuple, Dict, Any, Type, List
 
 
 class ProxyModule:
