+++ python/cuml/cuml/accel/magics.py
@@ -15,12 +15,7 @@
 #
 
 from cuml.accel.core import install
-from cuml.accel.magics import load_ipython_extension
-from cuml.accel.pytest_plugin import pytest_load_initial_conftests
 
 
-__all__ = (
-    "install",
-    "load_ipython_extension",
-    "pytest_load_initial_conftests",
-)
+def load_ipython_extension(ip):
+    install()
